update role and rule

+++ src/main/java/com/anlv/security/book/Book.java
@@ -48,9 +48,11 @@ public class Book {
             nullable = false,
             updatable = false
     )
-    private Integer createdBy;
+    private Long createdBy;
 
     @LastModifiedBy
     @Column(insertable = false)
-    private Integer lastModifiedBy;
+    private Long lastModifiedBy;
+
+
 }
